third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-length-percentage-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <length-percentage> with iterationComposite
  assert_equals: expected "calc(250% + 25px)" but got "calc(50% + 25px)"
Harness: the test ran to completion.
